+++ src/Controller/Backoffice/PostAdminController.php
@@ -8,6 +8,7 @@ use App\Controller\Frontoffice\SecurityController;
 use App\Controller\Frontoffice\UserController;
 use App\Model\Entity\Post;
 use App\Model\Repository\UserRepository;
+use App\Service\Authorization;
 use App\Service\Http\Request;
 use App\Service\Http\Response;
 use App\Service\Http\Session\Session;
@@ -35,7 +36,7 @@ final class PostAdminController
         $this->request = $request;
         $this->session = $session;
 
-        $security = new SecurityController($userRepository,$this->view,$this->session,$this->request);
+        $security = new Authorization($this->session,$this->request);
 
 
         if($security->notLogged() === true){
